NAS-136104: Service start/stop icon placement
