+++ src/app/pages/services/components/service-actions-cell/service-actions-cell.component.spec.ts
@@ -1,8 +1,6 @@
 import { TestbedHarnessEnvironment } from '@angular/cdk/testing/testbed';
 import { createHostFactory, mockProvider, SpectatorHost } from '@ngneat/spectator/jest';
 import { of } from 'rxjs';
-import { fakeSuccessfulJob } from 'app/core/testing/utils/fake-job.utils';
-import { mockApi, mockCall, mockJob } from 'app/core/testing/utils/mock-api.utils';
 import { mockAuth } from 'app/core/testing/utils/mock-auth.utils';
 import { ServiceName } from 'app/enums/service-name.enum';
 import { ServiceStatus } from 'app/enums/service-status.enum';
@@ -10,7 +8,6 @@ import { Service } from 'app/interfaces/service.interface';
 import { DialogService } from 'app/modules/dialog/dialog.service';
 import { IxIconHarness } from 'app/modules/ix-icon/ix-icon.harness';
 import { SlideIn } from 'app/modules/slide-ins/slide-in';
-import { ApiService } from 'app/modules/websocket/api.service';
 import { ServiceActionsCellComponent } from 'app/pages/services/components/service-actions-cell/service-actions-cell.component';
 import { ServiceFtpComponent } from 'app/pages/services/components/service-ftp/service-ftp.component';
 import { ServiceNfsComponent } from 'app/pages/services/components/service-nfs/service-nfs.component';
@@ -24,16 +21,11 @@ import { NvmeOfConfigurationComponent } from 'app/pages/sharing/nvme-of/nvme-of-
 describe('ServiceActionsCellComponent', () => {
   let spectator: SpectatorHost<ServiceActionsCellComponent>;
   let loader: ReturnType<typeof TestbedHarnessEnvironment.loader>;
-  let api: ApiService;
 
   const createHost = createHostFactory({
     component: ServiceActionsCellComponent,
     providers: [
       mockAuth(),
-      mockApi([
-        mockCall('service.update', 1),
-        mockJob('service.control', fakeSuccessfulJob()),
-      ]),
       mockProvider(SlideIn, {
         open: jest.fn(() => of()),
       }),
@@ -49,7 +41,6 @@ describe('ServiceActionsCellComponent', () => {
       hostProps: { service: service as Service },
     });
 
-    api = spectator.inject(ApiService);
     loader = TestbedHarnessEnvironment.loader(spectator.fixture);
   }
 
@@ -72,28 +63,6 @@ describe('ServiceActionsCellComponent', () => {
 
     const hasViewSessions = links.some((el) => el.textContent?.includes('View Sessions'));
     expect(hasViewSessions).toBe(true);
-  });
-
-  it('shows "Start Service" icon when service is stopped', async () => {
-    setup({ service: ServiceName.Ftp, state: ServiceStatus.Stopped });
-
-    const startIcon = await loader.getHarness(IxIconHarness.with({ name: 'mdi-play-circle' }));
-    expect(startIcon).toBeTruthy();
-
-    await startIcon.click();
-
-    expect(api.job).toHaveBeenCalledWith('service.control', ['START', ServiceName.Ftp, { silent: false }]);
-  });
-
-  it('shows "Stop Service" icon when service is running', async () => {
-    setup({ service: ServiceName.Ftp, state: ServiceStatus.Running });
-
-    const stopIcon = await loader.getHarness(IxIconHarness.with({ name: 'mdi-stop-circle' }));
-    expect(stopIcon).toBeTruthy();
-
-    await stopIcon.click();
-
-    expect(api.job).toHaveBeenCalledWith('service.control', ['STOP', ServiceName.Ftp]);
   });
 
   describe('edit', () => {
