+++ src/app/pages/services/components/service-actions-cell/service-actions-cell.component.ts
@@ -3,25 +3,14 @@ import {
 } from '@angular/core';
 import { MatIconButton } from '@angular/material/button';
 import { Router } from '@angular/router';
-import { UntilDestroy, untilDestroyed } from '@ngneat/until-destroy';
-import { TranslateModule, TranslateService } from '@ngx-translate/core';
-import {
-  filter, Observable, switchMap, take,
-} from 'rxjs';
+import { TranslateModule } from '@ngx-translate/core';
 import { RequiresRolesDirective } from 'app/directives/requires-roles/requires-roles.directive';
 import { AuditService } from 'app/enums/audit.enum';
-import { ServiceName, serviceNames, ServiceOperation } from 'app/enums/service-name.enum';
-import { ServiceStatus } from 'app/enums/service-status.enum';
-import { observeJob } from 'app/helpers/operators/observe-job.operator';
+import { ServiceName } from 'app/enums/service-name.enum';
 import { Service } from 'app/interfaces/service.interface';
-import { DialogService } from 'app/modules/dialog/dialog.service';
 import { IxIconComponent } from 'app/modules/ix-icon/ix-icon.component';
-import { LoaderService } from 'app/modules/loader/loader.service';
 import { SlideIn } from 'app/modules/slide-ins/slide-in';
-import { SnackbarService } from 'app/modules/snackbar/services/snackbar.service';
 import { TestDirective } from 'app/modules/test-id/test.directive';
-import { TranslatedString } from 'app/modules/translate/translate.helper';
-import { ApiService } from 'app/modules/websocket/api.service';
 import { ServiceFtpComponent } from 'app/pages/services/components/service-ftp/service-ftp.component';
 import { ServiceNfsComponent } from 'app/pages/services/components/service-nfs/service-nfs.component';
 import { ServiceSmbComponent } from 'app/pages/services/components/service-smb/service-smb.component';
@@ -30,12 +19,9 @@ import { ServiceSshComponent } from 'app/pages/services/components/service-ssh/s
 import { ServiceUpsComponent } from 'app/pages/services/components/service-ups/service-ups.component';
 import { GlobalTargetConfigurationComponent } from 'app/pages/sharing/iscsi/global-target-configuration/global-target-configuration.component';
 import { NvmeOfConfigurationComponent } from 'app/pages/sharing/nvme-of/nvme-of-configuration/nvme-of-configuration.component';
-import { ErrorHandlerService } from 'app/services/errors/error-handler.service';
-import { IscsiService } from 'app/services/iscsi.service';
 import { ServicesService } from 'app/services/services.service';
 import { UrlOptionsService } from 'app/services/url-options.service';
 
-@UntilDestroy()
 @Component({
   selector: 'ix-service-actions-cell',
   templateUrl: './service-actions-cell.component.html',
@@ -64,10 +50,6 @@ export class ServiceActionsCellComponent {
     return this.service().service === ServiceName.Cifs || this.service().service === ServiceName.Nfs;
   });
 
-  protected isRunning = computed(() => {
-    return this.service().state === ServiceStatus.Running;
-  });
-
   protected uniqueRowTag = computed(() => {
     return 'service-' + this.service().service.replace(/\./g, '');
   });
@@ -75,13 +57,6 @@ export class ServiceActionsCellComponent {
   constructor(
     private urlOptions: UrlOptionsService,
     private router: Router,
-    private dialogService: DialogService,
-    private translate: TranslateService,
-    private loader: LoaderService,
-    private api: ApiService,
-    private errorHandler: ErrorHandlerService,
-    private iscsiService: IscsiService,
-    private snackbar: SnackbarService,
     private servicesService: ServicesService,
     private slideIn: SlideIn,
   ) {}
@@ -92,25 +67,6 @@ export class ServiceActionsCellComponent {
 
   navigateToSessions(): void {
     this.router.navigate(this.sessionsUrl());
-  }
-
-  startService(): void {
-    this.api.job('service.control', [ServiceOperation.Start, this.service().service, { silent: false }]).pipe(
-      observeJob(),
-      this.loader.withLoader(),
-      this.errorHandler.withErrorHandler(),
-      untilDestroyed(this),
-    ).subscribe({
-      complete: () => this.snackbar.success(this.translate.instant('Service started')),
-    });
-  }
-
-  stopServiceClicked(): void {
-    this.confirmStop().pipe(
-      filter(Boolean),
-      take(1),
-      untilDestroyed(this),
-    ).subscribe(() => this.stopService());
   }
 
   configureService(): void {
@@ -142,52 +98,6 @@ export class ServiceActionsCellComponent {
       default:
         break;
     }
-  }
-
-  private confirmStop(): Observable<boolean> {
-    if (this.service().service === ServiceName.Iscsi) {
-      return this.confirmStopIscsiService();
-    }
-
-    return this.dialogService.confirm({
-      title: this.translate.instant('Confirm'),
-      message: this.translate.instant('Are you sure you want to stop {serviceName}?', { serviceName: serviceNames.get(this.service().service) }),
-      hideCheckbox: true,
-      buttonColor: 'warn',
-      buttonText: this.translate.instant('Stop'),
-    });
-  }
-
-  private stopService(): void {
-    this.api.job('service.control', [ServiceOperation.Stop, this.service().service]).pipe(
-      observeJob(),
-      this.loader.withLoader(),
-      this.errorHandler.withErrorHandler(),
-      untilDestroyed(this),
-    ).subscribe({
-      complete: () => this.snackbar.success(this.translate.instant('Service stopped')),
-    });
-  }
-
-  private confirmStopIscsiService(): Observable<boolean> {
-    return this.iscsiService.getGlobalSessions().pipe(
-      switchMap((sessions) => {
-        let message = this.translate.instant('Stop {serviceName}?', { serviceName: serviceNames.get(this.service().service) });
-        if (sessions.length) {
-          const connectionsMessage = this.translate.instant('{n, plural, one {There is an active iSCSI connection.} other {There are # active iSCSI connections}}', { n: sessions.length });
-          const stopMessage = this.translate.instant('Stop the {serviceName} service and close these connections?', { serviceName: serviceNames.get(this.service().service) });
-          message = `<font color="red">${connectionsMessage}</font><br>${stopMessage}` as TranslatedString;
-        }
-
-        return this.dialogService.confirm({
-          title: this.translate.instant('Alert'),
-          message,
-          hideCheckbox: true,
-          buttonText: this.translate.instant('Stop'),
-        });
-      }),
-      untilDestroyed(this),
-    );
   }
 
   private auditLogsUrl(): string {
