+++ src/app/pages/services/components/service-status-cell/service-status-cell.component.scss
@@ -1,15 +1,17 @@
 :host {
+  display: flex;
+  gap: 3px;
+}
+
+.status-pill {
+  border: 1px solid transparent;
+  border-radius: 20px;
   box-sizing: border-box;
   display: flex;
-
-  &.has-cell {
-    border: 1px solid transparent;
-    border-radius: 20px;
-    font-weight: bold;
-    gap: 6px;
-    max-width: max-content;
-    padding: 3px 8px;
-  }
+  font-weight: bold;
+  justify-content: center;
+  min-width: 72px;
+  padding: 3px 8px;
 
   &.running {
     border-color: var(--green);
